refactor: routes posts 모듈 분리
Related to: #143
